refactor(frontend): reorganiza estructura de componentes
- Mueve auth/ → components/auth/
- Mueve chart/ → components/charts/
- Mueve history/ → components/history/
- Actualiza imports en App.jsx y páginas
- Estructura más clara (7→4 carpetas en src/)

Simplifica navegación para perfiles junior
