+++ frontend/src/components/auth/AuthCard.jsx
