+++ frontend/src/components/auth/DataProtectionNotice.jsx
