+++ frontend/src/components/auth/UserPanelInfo.jsx
